Refactoring interfaces in a dedicated namespace
