+++ src/Paraunit/Parser/JSONLogParser.php
@@ -5,8 +5,8 @@ namespace Paraunit\Parser;
 use Paraunit\Exception\JSONLogNotFoundException;
 use Paraunit\Lifecycle\ProcessEvent;
 use Paraunit\Process\AbstractParaunitProcess;
-use Paraunit\TestResult\TestResultContainerInterface;
-use Paraunit\TestResult\TestResultInterface;
+use Paraunit\TestResult\Interfaces\TestResultContainerInterface;
+use Paraunit\TestResult\Interfaces\TestResultInterface;
 
 /**
  * Class JSONLogParser
