+++ src/Paraunit/Parser/RetryParser.php
@@ -4,8 +4,8 @@ namespace Paraunit\Parser;
 
 use Paraunit\Process\ProcessWithResultsInterface;
 use Paraunit\Process\RetryAwareInterface;
+use Paraunit\TestResult\Interfaces\TestResultContainerInterface;
 use Paraunit\TestResult\MuteTestResult;
-use Paraunit\TestResult\TestResultContainerInterface;
 use Paraunit\TestResult\TestResultFormat;
 
 /**
