+++ src/Paraunit/TestResult/Interfaces/FunctionNameInterface.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace Paraunit\TestResult;
+namespace Paraunit\TestResult\Interfaces;
 
 /**
  * Interface FunctionNameInterface
- * @package Paraunit\TestResult
+ * @package Paraunit\TestResult\Interfaces
  */
 interface FunctionNameInterface
 {
