+++ src/Paraunit/TestResult/Interfaces/TestResultInterface.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Paraunit\TestResult\Interfaces;
+
+/**
+ * Interface TestResultInterface
+ * @package Paraunit\TestResult\Interfaces
+ */
+interface TestResultInterface
+{
+}
