+++ src/Paraunit/TestResult/NullTestResult.php
@@ -2,6 +2,8 @@
 
 namespace Paraunit\TestResult;
 
+use Paraunit\TestResult\Interfaces\TestResultInterface;
+
 /**
  * Class NullTestResult
  * @package Paraunit\TestResult
