--- src/Paraunit/TestResult/TestResultInterface.php
@@ -1,11 +1,0 @@
-<?php
-
-namespace Paraunit\TestResult;
-
-/**
- * Interface TestResultInterface
- * @package Paraunit\Output\MuteTestResult
- */
-interface TestResultInterface
-{
-}
